The Discrete Uniform Distribution

The Discrete Uniform Distribution

ddunif(x, min = 0, max = 1)

pdunif(q, min = 0, max = 1)

qdunif(p, min = 0, max = 1)

rdunif(n, min = 0L, max = 1)

Arguments

x

vector of (non-negative integer) quantiles

min

minimum value of distribution (integer)

max

maximum value of distribution (integer)

q

vector of quantiles

p

vector of probabilities

n

number of random values to return

Value

ddunif gives the density, pdunif gives the distribution function, qdunif gives the quantile function, rdunif generates random deviates

Examples

ddunif(0:1)
#> [1] 0.5 0.5
pdunif(1)
#> [1] 1
qdunif(.5)
#> [1] 1
rdunif(10)
#>  [1] 1 1 0 1 1 1 0 1 1 0