Wrappers to compute goodness of fit test froms distfun objects

Wrappers to compute goodness of fit test froms distfun objects

ks_test(distfun, x, ...)

# S3 method for distfun
ad_test(distfun, x)

ad_test(distfun, x)

# S3 method for distfun
cvm_test(distfun, x)

cvm_test(distfun, x)

Arguments

distfun

a distfun object

x

numeric vector

...

arguments to be passed on to test function

Value

goodness of fit object

Examples

x <- rgamma(100, 1, 1)
fit <- fit_univariate(x, 'gamma')
ks_test(fit, x)
#> 
#> 	Asymptotic one-sample Kolmogorov-Smirnov test
#> 
#> data:  x
#> D = 0.060558, p-value = 0.8568
#> alternative hypothesis: two-sided
#> 
ad_test(fit, x)
#> 
#> 	Anderson-Darling test of goodness-of-fit
#> 	Null hypothesis: distribution 'distfun[[2]]'
#> 	Parameters assumed to be fixed
#> 
#> data:  x
#> An = 0.33754, p-value = 0.9073
#> 
cvm_test(fit, x)
#> 
#> 	Cramer-von Mises test of goodness-of-fit
#> 	Null hypothesis: distribution 'distfun[[2]]'
#> 	Parameters assumed to be fixed
#> 
#> data:  x
#> omega2 = 0.058529, p-value = 0.8244
#>